v2/requirements.txt
requests
python-keycloak
python-dotenv